Configure Firebase using React's Context API (following this instruction:https://www.robinwieruch.de/complete-firebase-authentication-react-tutorial#react-application-setup-create-react-app)
All Api keys are stored in .env then are called in firebase.js where the firebase is configured and initialized.

Current status: The UI of the sign in widget looks wrong. The Google
Account sign in button doesn't work, only sign in with Email works. When
signing in with Email, the getUserInfo function works ok: the user info
has been received and been able to be consoled. However, we don't know
how to render that info yet.